fix(ui): Fix splash theme issues.
