+++ app/src/main/res/values-night-v21/themes.xml
@@ -1,5 +1,14 @@
 <resources>
-    <style name="Theme.WorkSpace" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+    <style name="Theme.WorkSpace" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
+        <item name="android:windowBackground">@color/splash_background</item>
+    </style>
+
+    <style name="Theme.App.Splash" parent="Theme.SplashScreen">
+        <item name="windowSplashScreenAnimatedIcon">@android:color/transparent</item>
+        <item name="windowSplashScreenBackground">@color/splash_background</item>
+        <item name="windowSplashScreenIconBackgroundColor">@color/splash_background</item>
+        <item name="android:windowBackground">@color/splash_background</item>
+        <item name="postSplashScreenTheme">@style/Theme.WorkSpace</item>
     </style>
 </resources>
