+++ app/src/main/res/values-night/colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- Splash Screen Colors (Dark Mode) -->
+    <color name="splash_background">#1B1B22</color>
+</resources>
