+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources>
     <!-- Base application theme. -->
-    <style name="Theme.WorkSpace" parent="Theme.MaterialComponents.NoActionBar">
+    <style name="Theme.WorkSpace" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple_500</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
@@ -11,6 +11,7 @@
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
+        <item name="android:windowBackground">@color/splash_background</item>
         <!-- Customize your theme here. -->
     </style>
 
@@ -19,10 +20,11 @@
     </style>
 
     <style name="Theme.App.Splash" parent="Theme.SplashScreen">
-        <item name="windowSplashScreenAnimatedIcon">@drawable/avd_splash</item>
-        <item name="windowSplashScreenAnimationDuration">500</item>
-        <item name="windowSplashScreenBackground">#FFFFFF</item>
-
-        <item name="postSplashScreenTheme">@style/Theme.MaterialComponents.Light.NoActionBar</item>
+        <item name="windowSplashScreenAnimatedIcon">@android:color/transparent</item>
+        <item name="windowSplashScreenAnimationDuration">0</item>
+        <item name="windowSplashScreenBackground">@color/splash_background</item>
+        <item name="windowSplashScreenIconBackgroundColor">@color/splash_background</item>
+        <item name="android:windowBackground">@color/splash_background</item>
+        <item name="postSplashScreenTheme">@style/Theme.WorkSpace</item>
     </style>
 </resources>
