+++ data/src/main/java/com/flowintent/data/di/SecureModule.kt
@@ -8,6 +8,7 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.DataStoreFactory
 import androidx.datastore.dataStoreFile
+import com.flowintent.core.util.RootDetectionUtil
 import com.flowintent.data.secure.CryptoProvider
 import com.flowintent.data.secure.EncryptedSecurePrefsSerializer
 import com.flowintent.data.secure.SecurePrefs
