+++ gradle/libs.versions.toml
@@ -5,7 +5,6 @@ agp = "9.2.1"
 authKt = "3.4.1"
 coilCompose = "2.7.0"
 converterGson = "3.0.0"
-coreSplashscreen = "1.2.0"
 datastorePreferences = "1.2.1"
 desugar_jdk_libs = "2.1.5"
 detektGradlePlugin = "1.23.8"
@@ -16,6 +15,7 @@ hilt = "2.59.2"
 hiltNavigationCompose = "1.3.0"
 kotlin = "2.3.21"
 coreKtx = "1.18.0"
+coreSplashScreen = "1.0.1"
 junit = "4.13.2"
 junitVersion = "1.3.0"
 espressoCore = "3.7.0"
@@ -64,7 +64,7 @@ crashlytics = "3.0.7"
 accompanist-systemuicontroller = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "accompanistSystemuicontroller" }
 androidx-activity-compose = { module = "androidx.activity:activity-compose", version.ref = "activityCompose" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
-androidx-core-splashscreen = { module = "androidx.core:core-splashscreen", version.ref = "coreSplashscreen" }
+androidx-core-splashscreen = { group = "androidx.core", name = "core-splashscreen", version.ref = "coreSplashScreen" }
 androidx-datastore = { module = "androidx.datastore:datastore", version.ref = "datastorePreferences" }
 androidx-datastore-preferences = { module = "androidx.datastore:datastore-preferences", version.ref = "datastorePreferences" }
 androidx-hilt-navigation-compose = { module = "androidx.hilt:hilt-navigation-compose", version.ref = "hiltNavigationCompose" }
